Create a new top-level export function in the library.
This is just the body of the old runCommand function from the export
tool, moved out on its own and with all the imports updated.
